Simplify enum generation and fix clang not adding enums
The functions for generating the beginning and ending of an enum are now
fused to a single function. Users can now call
EnumEntity::generateValues to generate the enum entries.

The value of an enum entry is now a string instead of an integer as the
value format can vary between languages.

Additionally the clang backend now exports enums to the hierarchy.