Make FunctionEntity more dependent on FunctionGroupEntity
Since FunctionGroupEntity represents a group of function overloads, they
should be of the same type and name. These details are now stored in
FunctionGroupEntity instead of FunctionEntity.

Having a guarantee that FunctionEntities reside inside a FunctionGroupEntity
makes resolving much more nice, since when you want to locate a function
which may or may not have multiple overloads, you can simply assume that
the result of the resolve() call is a FunctionGroupEntity holding the
overloads that you might want.